Adding support for assignment for chained calls.
Not fully correct, because parser accepts assignment to a function call, but it is not easily possible to create such grammar using descent recursive parser or Sprache. Easier is custom parsing step that makes sure that last chain call is just a member access not a function call. Even C# seems to use this workaround.