Bringing back shadow functions for defnames and transforming index member access from `[idx]` to `_idx_`.
The reason is that defname value either has be prefixed with `def.` or must be accessed from `<eval`, not from macro directly. So transpiler needs to know if particular symbol is function name or defname to do proper transpilation.
But there are function/defnames conflicts that can be resolved only at runtime ONLY like `attr_invis` and `attr_decay` (both native functions and defnames). Even if the name is transformed and, user can still declare both defname and function with the same name creating the same conflict.

Function name can't contain `[]`, so brackets needs to be transformed to `_`.

Most likely it is possible to distinguish between defname access and function call at build time somehow but I don't see it at this moment. Shadow functions are easy way to move forward immediately.