Bit shits (sic!) operators.
These operators are ambiguous on both .99 and .56 versions in different ways. I cannot figure out how to parse `>>` properly without spaces around it on .99. On .56 version this operator doesn't work at all because interpreter confuses it with macro ending. The only workaround is to replace `>>` by macro expansion (`op_shiftright`) and no, this workaround doesn't work when `op_shiftright` is defname.