Pre-transpilation source code replacement implemented.
There are source code parts that cannot be transpiled properly like code in literals. It is handy to be able to process this code as well. Source code base is supposed to contain only limited number of these cases. An alternative could be some kind of annotations directly in source code (comments?). There is no automated way how to solve it in general.
Replacements hardcoded for Erebor, if this feature becomes more useful it has to be moved to some "project" file which could contain more information like list of files being transpiled.